Integrate notifications component
